replace MonadUnliftIO with MonadBaseControl IO
The movitation for this change comes mainly from using morpheus with the
snap websever framework. While the `snap` monad has a `MonadBaseControl
IO` instance, it lacks a `MonadUnliftIO` instance. I believe the
`MonadBaseControl IO` is more general than `MonadUnliftIO` and will help
for other computational contexts as well.

In additon, the ping thread is removed from the websocket, because again
webservers like `snap` will add a configurable ping thread when invoking
their websocket handler, so there is no need for doing this in the
morpheus library.